requirements.txt
amqp==2.5.2
asgiref==3.2.5
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-oauth-toolkit==1.3.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
